Break find_components.dart into 3 files: -> find_components.dart which defined the component visitor. -> component_visitor_exceptions.dart which defines the exceptions and the error handling code. -> lifecycle_hooks.dart which contains the LifecycleHook extractor. This breaks a circular import chain between find_components.dart and visit_pipes.dart
PiperOrigin-RevId: 229804254
