+++ _tests/test/resolve_util.dart
@@ -3,6 +3,7 @@ import 'dart:async';
 import 'package:analyzer/dart/element/element.dart';
 import 'package:angular/src/compiler/offline_compiler.dart';
 import 'package:angular/src/source_gen/template_compiler/find_components.dart';
+import 'package:angular/src/source_gen/template_compiler/component_visitor_exceptions.dart';
 import 'package:build/build.dart';
 import 'package:build_test/build_test.dart';
 import 'package:source_gen/source_gen.dart';
@@ -20,6 +21,6 @@ Future<NormalizedComponentWithViewDirectives> resolveAndFindComponent(
   final library = await resolve("import 'package:angular/angular.dart';"
       "$source");
   final artifacts = findComponentsAndDirectives(
-      LibraryReader(library), FindComponentsExceptionHandler());
+      LibraryReader(library), ComponentVisitorExceptionHandler());
   return artifacts.components.first;
 }
