+++ angular/lib/src/source_gen/template_compiler/find_components.dart
@@ -3,12 +3,9 @@ import 'package:analyzer/dart/constant/value.dart';
 import 'package:analyzer/dart/element/element.dart';
 import 'package:analyzer/dart/element/type.dart';
 import 'package:analyzer/dart/element/visitor.dart';
-import 'package:analyzer/error/error.dart';
-import 'package:analyzer/src/dart/analysis/results.dart';
 import 'package:analyzer/src/dart/element/element.dart';
 import 'package:build/build.dart';
 import 'package:source_gen/source_gen.dart';
-import 'package:source_span/source_span.dart';
 import 'package:angular/src/compiler/analyzed_class.dart';
 import 'package:angular/src/compiler/compile_metadata.dart';
 import 'package:angular/src/compiler/expression_parser/ast.dart' as ast;
@@ -23,10 +20,11 @@ import 'package:angular/src/core/metadata/lifecycle_hooks.dart';
 import 'package:angular/src/source_gen/common/annotation_matcher.dart';
 import 'package:angular/src/source_gen/common/url_resolver.dart';
 import 'package:angular_compiler/angular_compiler.dart';
-import 'package:angular_compiler/cli.dart';
 
 import 'compile_metadata.dart';
+import 'component_visitor_exceptions.dart';
 import 'dart_object_utils.dart';
+import 'lifecycle_hooks.dart';
 import 'pipe_visitor.dart';
 
 const String _visibilityProperty = 'visibility';
@@ -35,7 +33,7 @@ const _statefulDirectiveFields = [
 ];
 
 AngularArtifacts findComponentsAndDirectives(
-    LibraryReader library, FindComponentsExceptionHandler exceptionHandler) {
+    LibraryReader library, ComponentVisitorExceptionHandler exceptionHandler) {
   var componentVisitor = _NormalizedComponentVisitor(library, exceptionHandler);
   library.element.accept(componentVisitor);
   return AngularArtifacts(
@@ -49,7 +47,7 @@ class _NormalizedComponentVisitor extends RecursiveElementVisitor<Null> {
   final List<CompileDirectiveMetadata> directives = [];
   final LibraryReader _library;
 
-  final FindComponentsExceptionHandler _exceptionHandler;
+  final ComponentVisitorExceptionHandler _exceptionHandler;
 
   _NormalizedComponentVisitor(this._library, this._exceptionHandler);
 
@@ -183,7 +181,7 @@ class _ComponentVisitor
   final _viewQueries = <CompileQueryMetadata>[];
 
   final LibraryReader _library;
-  final FindComponentsExceptionHandler _exceptionHandler;
+  final ComponentVisitorExceptionHandler _exceptionHandler;
 
   /// Whether the component being visited re-implements 'noSuchMethod'.
   bool _implementsNoSuchMethod = false;
@@ -750,31 +748,12 @@ class _ComponentVisitor
   }
 }
 
-List<LifecycleHooks> extractLifecycleHooks(ClassElement clazz) {
-  const hooks = <TypeChecker, LifecycleHooks>{
-    TypeChecker.fromRuntime(OnInit): LifecycleHooks.onInit,
-    TypeChecker.fromRuntime(OnDestroy): LifecycleHooks.onDestroy,
-    TypeChecker.fromRuntime(DoCheck): LifecycleHooks.doCheck,
-    TypeChecker.fromRuntime(OnChanges): LifecycleHooks.onChanges,
-    TypeChecker.fromRuntime(AfterChanges): LifecycleHooks.afterChanges,
-    TypeChecker.fromRuntime(AfterContentInit): LifecycleHooks.afterContentInit,
-    TypeChecker.fromRuntime(AfterContentChecked):
-        LifecycleHooks.afterContentChecked,
-    TypeChecker.fromRuntime(AfterViewInit): LifecycleHooks.afterViewInit,
-    TypeChecker.fromRuntime(AfterViewChecked): LifecycleHooks.afterViewChecked,
-  };
-  return hooks.keys
-      .where((hook) => hook.isAssignableFrom(clazz))
-      .map((t) => hooks[t])
-      .toList();
-}
-
 /// Ensures that all entries in [directiveTypes] match an entry in [directives].
 void _errorOnUnusedDirectiveTypes(
     ClassElement element,
     List<CompileDirectiveMetadata> directives,
     List<CompileTypedMetadata> directiveTypes,
-    FindComponentsExceptionHandler exceptionHandler) {
+    ComponentVisitorExceptionHandler exceptionHandler) {
   if (directiveTypes.isEmpty) return;
 
   // Creates a unique key given a module URL and symbol name.
@@ -790,209 +769,6 @@ void _errorOnUnusedDirectiveTypes(
       exceptionHandler.handle(UnusedDirectiveTypeError(element, directiveType));
     }
   }
-}
-
-class IndexedAnnotation<T extends Element> {
-  final T element;
-  final ElementAnnotation annotation;
-  final int annotationIndex;
-
-  IndexedAnnotation(this.element, this.annotation, this.annotationIndex);
-}
-
-class FindComponentsExceptionHandler {
-  final List<AsyncBuildError> _errors = [];
-
-  void handle(AsyncBuildError error) {
-    _errors.add(error);
-  }
-
-  Future<void> maybeReportErrors() async {
-    if (_errors.isEmpty) {
-      return;
-    }
-    final buildErrors =
-        await Future.wait(_errors.map((error) => error.resolve()));
-    throw BuildError(
-        buildErrors.map((buildError) => buildError.message).join("\n\n"));
-  }
-}
-
-class AsyncBuildError extends BuildError {
-  Future<BuildError> resolve() => Future.value(this);
-}
-
-class AngularAnalysisError extends AsyncBuildError {
-  final List<AnalysisError> constantEvaluationErrors;
-  final IndexedAnnotation<ClassElement> indexedAnnotation;
-
-  AngularAnalysisError(this.constantEvaluationErrors, this.indexedAnnotation);
-
-  @override
-  Future<BuildError> resolve() async {
-    bool hasOffsetInformation =
-        constantEvaluationErrors.any((error) => error.offset != 0);
-    // If this code is called from a tool using [AnalysisResolvers], then
-    //   1) [constantEvaluationErrors] already has source location information
-    //   2) [ResolvedLibraryResultImpl] will NOT have the errors.
-    // So, we return immediately with the information in
-    // [constantEvaluationErrors].
-    if (hasOffsetInformation) {
-      return Future.value(_buildErrorForAnalysisErrors(
-          constantEvaluationErrors, indexedAnnotation.element));
-    }
-
-    var libraryElement = indexedAnnotation.element.library;
-    var libraryResult = await ResolvedLibraryResultImpl.tmp(libraryElement);
-    var classResult =
-        libraryResult.getElementDeclaration(indexedAnnotation.element);
-
-    ClassDeclaration classDeclaration = classResult.node;
-    final resolvedAnnotation =
-        classDeclaration.metadata[indexedAnnotation.annotationIndex];
-
-    // Only include the errors that are inside the annotation.
-    return _buildErrorForAnalysisErrors(
-        classResult.resolvedUnit.errors.where((error) =>
-            error.offset >= resolvedAnnotation.offset &&
-            error.offset <= resolvedAnnotation.end),
-        indexedAnnotation.element);
-  }
-
-  BuildError _buildErrorForAnalysisErrors(
-      Iterable<AnalysisError> errors, ClassElement componentType) {
-    return BuildError(
-      messages.unresolvedSource(
-        errors.map((e) {
-          final sourceUrl = e.source.uri;
-          final sourceContent = e.source.contents.data;
-
-          if (sourceContent.isEmpty) {
-            return SourceSpanMessageTuple(
-                SourceSpan(
-                    SourceLocation(0, sourceUrl: sourceUrl, line: 0, column: 0),
-                    SourceLocation(0, sourceUrl: sourceUrl, line: 0, column: 0),
-                    ''),
-                '${e.message} [with offsets into source file missing]');
-          }
-
-          return SourceSpanMessageTuple(
-            sourceSpanWithLineInfo(
-              e.offset,
-              e.length,
-              sourceContent,
-              sourceUrl,
-            ),
-            e.message,
-          );
-        }),
-        reason: ''
-            'Compiling @Component-annotated class "${componentType.name}" '
-            'failed.\n\n${messages.analysisFailureReasons}',
-      ),
-    );
-  }
-}
-
-class UnresolvedExpressionError extends AsyncBuildError {
-  final Iterable<AstNode> expressions;
-  final ClassElement componentType;
-  final CompilationUnitElement compilationUnit;
-
-  UnresolvedExpressionError(
-      this.expressions, this.componentType, this.compilationUnit);
-
-  @override
-  Future<BuildError> resolve() =>
-      Future.value(_buildErrorForUnresolvedExpressions(
-          expressions, componentType, compilationUnit));
-
-  // TODO(deboer): Since we are checking ElementAnnotation.constantValueErrors,
-  // all code paths that call this function are unreachable.
-  // If we don't see any errors in the wild, delete this code.
-  BuildError _buildErrorForUnresolvedExpressions(Iterable<AstNode> expressions,
-      ClassElement componentType, CompilationUnitElement compilationUnit) {
-    return BuildError(
-      messages.unresolvedSource(
-        expressions.map((e) {
-          return SourceSpanMessageTuple(
-            sourceSpanWithLineInfo(
-              e.offset,
-              e.length,
-              componentType.source.contents.data,
-              componentType.source.uri,
-              lineInfo: compilationUnit.lineInfo,
-            ),
-            'This argument *may* have not been resolved',
-          );
-        }),
-        reason: ''
-            'Compiling @Component annotated class "${componentType.name}" '
-            'failed.\n'
-            'NOTE: Your build triggered an error in the Angular error reporting\n'
-            'code. Please report a bug: ${messages.urlFileBugs}\n'
-            '\n\n${messages.analysisFailureReasons}',
-      ),
-    );
-  }
-}
-
-class UnusedDirectiveTypeError extends ErrorMessageForAnnotation {
-  final ClassElement element;
-  final CompileTypedMetadata directiveType;
-
-  static IndexedAnnotation firstComponentAnnotation(ClassElement element) {
-    final index = element.metadata.indexWhere(isComponent);
-    if (index == -1)
-      throw ArgumentError("[element] must have a @Component annotation");
-    return IndexedAnnotation(element, element.metadata[index], index);
-  }
-
-  UnusedDirectiveTypeError(this.element, this.directiveType)
-      : super(
-            firstComponentAnnotation(element),
-            'Entry in "directiveTypes" missing corresponding entry in '
-            '"directives" for "${directiveType.name}".\n\n'
-            'If you recently removed "${directiveType.name}" from "directives", '
-            'please also remove its corresponding entry from "directiveTypes".');
-}
-
-class ErrorMessageForAnnotation extends AsyncBuildError {
-  final IndexedAnnotation indexedAnnotation;
-  final String message;
-
-  ErrorMessageForAnnotation(this.indexedAnnotation, this.message);
-
-  @override
-  Future<BuildError> resolve() async {
-    final element = indexedAnnotation.element;
-    final annotationIndex = indexedAnnotation.annotationIndex;
-
-    var libraryResult = await ResolvedLibraryResultImpl.tmp(element.library);
-    var result = libraryResult.getElementDeclaration(element);
-    AnnotatedNode annotatedNode = result.node;
-    if (annotatedNode.metadata.length != element.metadata.length) {
-      // [ResolvedLibraryResultImpl] lost the metadata, so return the
-      // original annotation even though it probably doesn't have the correct
-      // information.
-      return BuildError.forAnnotation(
-          element.metadata[annotationIndex], message);
-    }
-    ElementAnnotation resolvedAnnotation =
-        annotatedNode.metadata[annotationIndex].elementAnnotation;
-    return BuildError.forAnnotation(resolvedAnnotation, message);
-  }
-}
-
-class ErrorMessageForElement extends AsyncBuildError {
-  final Element element;
-  final String message;
-
-  ErrorMessageForElement(this.element, this.message);
-
-  @override
-  Future<BuildError> resolve() =>
-      Future.value(BuildError.forElement(element, message));
 }
 
 void _prohibitBindingChange(
